Added the About and Contact Pages
About and Contact pages implemented. Minor tweaks to the media scroller element to improve readability.

I had to use transform:translateY to get all of the cards lined up the way I wanted to. this is a bandaid fix at best but if I ever come back to it I'll do something more robust.